+++ enums.js
@@ -1,0 +1,12 @@
+module.exports.SyncStatus = {
+    COMPLETE: "complete",
+    ALREADY_DONE: "already done",
+    NO_DATA: "no data",
+    FAILED: "failed",
+    UPDATING: "updating",
+    HISTORY: "history",
+    ACCOUNT_INACTIVE: "account-inactive",
+    NOT_SYNCING: "",
+}
+
+Object.freeze(module.exports.SyncStatus);